backend/requirements.txt
boto3==1.34.144
requests==2.31.0
websocket-client==1.6.4
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
matplotlib==3.7.2